requirements.txt
fastapi
uvicorn
pymongo
requests
pyjwt
python-multipart

# testing
pytest
coverage
aiounittest
